Deviation had been changed

+++ MyWebApi/MyWebApi/Entities/UserInfoEntities/User.cs
@@ -22,7 +22,7 @@ namespace MyWebApi.Entities.UserInfoEntities
         public bool? HasPremium { get; set; }
         public bool? HadReceivedReward { get; set; }
         public short? DailyRewardPoint { get; set; }
-        public short? BonusIndex { get; set; }
+        public double? BonusIndex { get; set; }
         public string? Nickname { get; set; }
         public long? ParentId { get; set; }
         public int? ProfileViewsCount { get; set; }
